goal: goal table generator for water
A new generator allows the previous tables with the callback functions to
be automatically generated for simple targets. It removes redundancy (not the
same string in many level files) and improves readability. The technique used
is still the same and programmers can still supply their own tables, but for
these common goals the generator should be used.