level/tui: employ a simple method to prevent level files from accessing globals
A level was sourced into the code until now, hoping that the level does not
break anything and provides a level variable to the caller. Until now. From
now on, a simple method is employed to prevent it from overwriting globals
accidentally (it can still use _G.foo = ... to fuck things up, but this is
irrelevant now). The new wrapper function creates a new environment and returns
whatever level variable was set in the new environment... the calling function
(something in the interface code, that is) then assigns this to its own level
variable.