converted the project to Perl
The thing that most annoys me of Lua is that it is VERY incapable. The
community is about non-existant and there are almost no modules that one
can install. You have to write EVERYTHING your ducking self. Sure, that
IS easy, but it still takes time and actually, it's a bit more complex to
make it look better than a randomly snot level.

Lua definitely has the nicer syntax, at the very least for level files.
That definitely does not help me if I can't implement frontends for ncurses
nor even (what I personally prefer) doing it myself, because there are simply
no bindings to termios functions (tc*)! (There are two POSIX modules for Lua
but they're horribly incomplete.)

The only remaining choice was therefore to rewrite the project in Perl
completely, hoping that new frontends can be implemented now.